Actually got pastoria working!
