+++ packages/pastoria-runtime/src/index.ts
@@ -1,5 +1,8 @@
 export {
+  AnyPreloadedEntryPoint,
+  AnyPreloadedQuery,
+  EnvironmentProvider,
   relayClientEnvironment,
   RelayResolverContext,
-  AnyPreloadedEntryPoint,
+  RouterOps,
 } from './relay_client_environment';
