+++ packages/pastoria/templates/router.tsx
@@ -1,4 +1,9 @@
-import {relayClientEnvironment, AnyPreloadedEntryPoint} from 'pastoria-runtime';
+import {
+  AnyPreloadedEntryPoint,
+  EnvironmentProvider,
+  relayClientEnvironment,
+  RouterOps,
+} from 'pastoria-runtime';
 import {createRouter} from 'radix3';
 import {
   AnchorHTMLAttributes,
@@ -16,17 +21,11 @@ import {preinit, preloadModule} from 'react-dom';
 import {
   EntryPoint,
   EntryPointContainer,
-  EnvironmentProviderOptions,
-  IEnvironmentProvider,
   loadEntryPoint,
   RelayEnvironmentProvider,
   useEntryPointLoader,
 } from 'react-relay/hooks';
-import {OperationDescriptor, PayloadData} from 'relay-runtime';
-import type {Manifest} from 'vite';
 import * as z from 'zod/v4-mini';
-
-export type RouterOps = [OperationDescriptor, PayloadData][];
 
 type RouterConf = typeof ROUTER_CONF;
 const ROUTER_CONF = {
@@ -123,9 +122,7 @@ function useLocation(initialPath?: string) {
   return [location, setLocation] as const;
 }
 
-export function router__hydrateStore(
-  provider: IEnvironmentProvider<EnvironmentProviderOptions>,
-) {
+export function router__hydrateStore(provider: EnvironmentProvider) {
   const env = provider.getEnvironment(null);
   if ('__router_ops' in window) {
     const ops = (window as any).__router_ops as RouterOps;
@@ -136,7 +133,7 @@ export function router__hydrateStore(
 }
 
 export async function router__loadEntryPoint(
-  provider: IEnvironmentProvider<EnvironmentProviderOptions>,
+  provider: EnvironmentProvider,
   initialPath?: string,
 ) {
   if (!initialPath) initialPath = window.location.href;
@@ -161,13 +158,6 @@ const RouterContext = createContext<RouterContextValue>({
   setLocation: () => {},
 });
 
-export interface RouterBootstrap {
-  preloadModules: string[];
-  preloadStylesheets: string[];
-  bootstrapScriptContent: string;
-  bootstrapModules: string[];
-}
-
 const REACT_REFRESH_SCRIPT = `
 import RefreshRuntime from 'http://localhost:3000/@react-refresh'
 RefreshRuntime.injectIntoGlobalHook(window)
@@ -177,7 +167,7 @@ window.__vite_plugin_react_preamble_installed__ = true`;
 
 export function router__createAppFromEntryPoint(
   initialEntryPoint: AnyPreloadedEntryPoint | null,
-  provider: IEnvironmentProvider<EnvironmentProviderOptions>,
+  provider: EnvironmentProvider,
   initialPath?: string,
 ) {
   const env = provider.getEnvironment(null);
@@ -276,7 +266,7 @@ export function router__createAppFromEntryPoint(
   function RouterApp(props: {
     preloadModules?: string[];
     preloadStylesheets?: string[];
-    App?: React.ComponentType<PropsWithChildren<{}>>;
+    App?: React.ComponentType<PropsWithChildren<{}>> | null;
   }) {
     return (
       <RouterShell
@@ -291,13 +281,6 @@ export function router__createAppFromEntryPoint(
       </RouterShell>
     );
   }
-
-  RouterApp.bootstrap = (manifest?: Manifest): RouterBootstrap => ({
-    preloadModules: [],
-    preloadStylesheets: [],
-    bootstrapScriptContent: '',
-    bootstrapModules: [],
-  });
 
   return RouterApp;
 }
