+++ src/lib/server/context.ts
@@ -1,9 +1,8 @@
 import DataLoader from 'dataloader';
-import express from 'express';
 
 /** @gqlContext */
 export class Context {
-  constructor(private readonly request: express.Request) {}
+  constructor() {}
 
   private readonly DERIVED_CACHE = new Map<string, unknown>();
